Add background for playscene
